feat: move hooks together with existing hooks package
